% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCohorts.r
\name{createCohorts}
\alias{createCohorts}
\title{Generate cohort}
\usage{
createCohorts(
  connection,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  targetCohortTable,
  createTargetCohortTable = FALSE,
  conditionConceptIds,
  measurementConceptId,
  observationWindowBefore,
  observationWindowAfter,
  drugWindow,
  targetCohortId
)
}
\arguments{
\item{connection}{Name of local folder where the results were generated; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{cdmDatabaseSchema}{How many parallel cores should be used? If more cores are made
available this can speed up the analyses.}
}
\description{
Generate cohort
}
\details{
This function generates cohort for the medical deconfounder (single outcome)
}
