% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{generateData}
\alias{generateData}
\title{Listing function}
\usage{
generateData(
  connection,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  targetCohortTable,
  drugExposureTable,
  measurementTable,
  conditionConceptIds,
  measurementConceptId,
  observationWindowBefore,
  observationWindowAfter,
  drugWindow,
  createTargetCohortTable = T,
  createTargetCohort = T,
  extractFeature = T,
  targetCohortId = NULL,
  dataFolder,
  drugFilename,
  measFilename
)
}
\description{
Listing function
}
