% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{generateMvdData}
\alias{generateMvdData}
\title{Listing function}
\usage{
generateMvdData(
  connection,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  targetCohortTable,
  minimumProportion = 0.001,
  targetDrugTable = "DRUG_ERA",
  ingredientConceptIds = NULL,
  measurementConceptIds = NULL,
  createTargetCohort = T,
  extractDrugFeature = T,
  extractMeasFeature = T,
  labWindow = 35,
  targetCohortId = NULL,
  temporalStartDays = c(-35, 1),
  temporalEndDays = c(-1, 35),
  sampleSize = NULL,
  outputFolder
)
}
\description{
Listing function
}
