% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listingIngredients.R
\name{listingIngredients}
\alias{listingIngredients}
\title{Generate list of concept Ids of interest}
\usage{
listingIngredients(
  connection,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  minimumProportion = 0.01,
  targetDrugTable = "DRUG_ERA"
)
}
\arguments{
\item{connection}{Name of local folder where the results were generated; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{cdmDatabaseSchema}{How many parallel cores should be used? If more cores are made
available this can speed up the analyses.}
}
\description{
Generate list of concept Ids of interest
}
\details{
This function
}
