% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{unadjIndeptRidgeReg}
\alias{unadjIndeptRidgeReg}
\title{Listing function}
\usage{
unadjIndeptRidgeReg(
  inputFolder,
  outputFolder,
  cvFold = 5,
  lambdas = 10^seq(2, -2, by = -0.1)
)
}
\description{
Listing function
}
