% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCohorts.r
\name{createCohorts}
\alias{createCohorts}
\title{Generate cohort}
\usage{
createCohorts(
  connection,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  targetCohortTable,
  createTargetCohortTable = TRUE,
  conditionConceptIds,
  measurementConceptId,
  observationWindowBefore,
  observationWindowAfter,
  drugWindow,
  targetCohortId
)
}
\arguments{
\item{connection}{Connection to the database. An object of class connect as created by the connect function in the DatabaseConnector package.}

\item{cdmDatabaseSchema}{A schema where OMOP CDM data are stored.}

\item{oracleTempSchema}{A schema that can be used to create temp tables in when using Oracle.}

\item{vocabularyDatabaseSchema}{A schema where vocabulary is stored}

\item{cohortDatabaseSchema}{A schema where the cohort is stored}

\item{targetCohortTable}{A string corresponds to the name of cohort table}

\item{createTargetCohortTable}{A boolean that indicates whether the targetCohortTable will be created. Default TRUE.}

\item{conditionConceptIds}{A list of condition concept IDs that correspond to the disease of interest.}

\item{measurementConceptId}{A numeric of measurement concept ID}

\item{observationWindowBefore}{An integer indicates the number of pre-treatment days used to look for pre-treatment measurement.}

\item{observationWindowAfter}{An integer indicates the number of post-treatment days used to look for post-treatment measurement.}

\item{drugWindow}{An integer indicates the number of post-treatment days during which drug exposure are also considered. Default is 0}

\item{targetCohortId}{An integer of the cohort ID.}
}
\value{
Returns a string containing the rendered SQL.
}
\description{
Generate cohort
}
\details{
This function generates cohort for the medical deconfounder (single outcome)
}
