% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCohort.r
\name{createCohorts}
\alias{createCohorts}
\title{Generate cohort}
\usage{
createCohorts(connection, cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema, cohortDatabaseSchema,
  cohortTable, oracleTempSchema, ingredientConceptIds,
  measurementConceptIds, drugWindow = 35, labWindow = 35,
  targetCohortId, outputFolder)
}
\arguments{
\item{connection}{Name of local folder where the results were generated; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{cdmDatabaseSchema}{How many parallel cores should be used? If more cores are made
available this can speed up the analyses.}
}
\description{
Generate cohort
}
\details{
This function generates cohort for the multivariate deconfounder
}
