% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{fitDeconfounder}
\alias{fitDeconfounder}
\title{Listing function}
\usage{
fitDeconfounder(
  learning_rate,
  max_steps,
  latent_dim,
  batch_size,
  num_samples,
  holdout_portion,
  print_steps,
  tolerance,
  num_confounder_samples,
  cv,
  outcome_type,
  project_dir
)
}
\description{
Listing function
}
